Changed HexEditorPrivate to use float precision for character sizes. This should fix some errors with special fonts like on Mac OS X.

+++ src/gtatools-gui/src/src/gui/HexEditorPrivate.h
@@ -128,7 +128,7 @@ private:
 private:
 	int enabledParts;
 	QTimer cursorBlinkTimer;
-	int charW, charH, charAsc, charDesc;
+	float charW, charH, charAsc, charDesc;
 	QByteArray data;
 	int addrHexGapSize;
 	int hexAsciiGapSize;
